backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
PyGithub==2.2.0
python-jenkins==1.8.0
openai==1.12.0